Add command line arguments for launching polybot
