+++ gradle.properties
@@ -3,7 +3,7 @@
 # Copyright (c) 2021-2021 solonovamax <solonovamax@12oclockpoint.com>
 #
 # The file gradle.properties is part of PolyhedralBot
-# Last modified on 31-12-2021 11:15 p.m.
+# Last modified on 31-12-2021 11:46 p.m.
 #
 # MIT License
 #
@@ -47,6 +47,7 @@ KOTLINX_SERIALIZATION_VERSION=1.3.2
 KOTLINX_COROUTINES_VERSION=1.6.0-native-mt
 KOTLINX_UUID_VERSION=0.0.12
 JETBRAINS_ANNOTATIONS_VERSION=22.0.0
+CLIKT_VERSION=3.3.0
 JDA_VERSION=4.3.0_313
 DISCORD_WEBHOOKS_VERSION=0.7.4
 JDA_KTX_VERSION=0.7.0
